+++ Hazel/src/Hazel/Renderer/Buffer.h
@@ -142,7 +142,7 @@ namespace Hazel {
 		virtual void Unbind() const  = 0;
 
 		virtual uint32_t GetCount() const = 0;
-		static Ref<IndexBuffer> Create(uint32_t* indices, uint32_t size);
+		static Ref<IndexBuffer> Create(uint16_t* indices, uint32_t size);
 
 	};
